pyproject.toml
[project]
name = "sps-genai"
version = "0.4.0"
description = "Assignment 4: Diffusion and Energy-Based Models for Image Generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic>=2.0.0",
    "spacy>=3.7",
    "numpy>=1.24.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.66.0",
]
